requirements.txt
# Core
Django==4.1.5
django-configurations==2.4
gunicorn==20.1.0

# For the persistence stores
psycopg2-binary==2.9.5
dj-database-url==1.2.0

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
django-filter==22.1
drf-spectacular==0.24.0

# Static and Media Storage
django-storages==1.13.2

# Environments
python-dotenv==0.21.0